+++ R/PowerLaw.R
@@ -26,16 +26,14 @@ fitPlaw <- function(graph) {
   probability <- probability[nonzero.position]
   degree <- degree[nonzero.position]
   reg <- lm(log(probability) ~ log(degree))
-  cozf <- coef(reg)
-  power.law.fit <- function(x) exp(cozf[[1]] + cozf[[2]] * log(x))
+  cf <- coef(reg)
+  plotnet <- data.frame(probability = probability, degree = degree)
   message(paste("Alpha =",
                 round(-cozf[[2]], 3)))
   message(paste("R square =",
                 round(summary(reg)$r.squared, 3)))
-  # plot
-  plot(probability ~ degree, log = "xy",
-       xlab = "Degree (log)", ylab = "Probability (log)",
-       col = 1, main = "Degree Distribution",
-       pch = 19, cex = .5)
-  curve(power.law.fit, col = "red", add = TRUE, n = length(d))
+  ggplot(plotnet, aes(y=log(probability), x=log(degree))) +
+    geom_point(col = "#56B4E9")+
+    geom_smooth(method="lm", se=FALSE, col = "#E69F00")
+
 }